reorganize app structure, implemented global settings
