+++ src/main/java/misc/ConcurrentCustomerLine.java
@@ -5,16 +5,20 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Random;
 import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import person.AbstractCustomer;
 import person.Patient;
-import setup.CustomerSupplier;
-import setup.PatientSupplier;
+import setup.AppConfig;
+import utils.CustomerSupplier;
+import utils.PatientSupplier;
 
 public class ConcurrentCustomerLine extends ConcurrentLinkedQueue<AbstractCustomer> {
 
     private static final Logger LOG = LogManager.getLogger(ConcurrentCustomerLine.class);
+    private Lock dequeueLock = new ReentrantLock();
 
     public void addCustomer(AbstractCustomer abstractCustomer) {
         LOG.trace(abstractCustomer.getName() + " is now in line");
@@ -34,14 +38,20 @@ public class ConcurrentCustomerLine extends ConcurrentLinkedQueue<AbstractCustom
     }
 
     public AbstractCustomer getNextCustomer() {
-        AbstractCustomer customer = poll();
-        if (customer == null) {
-            LOG.warn("The line is empty" + " from thread "
-                + Thread.currentThread().getName());
-        } else {
-            LOG.trace("The next customer is: " + customer.getName());
+        dequeueLock.lock();
+        try{
+            AbstractCustomer customer = poll();
+            if (customer == null) {
+                LOG.warn("The line is empty" + " from thread "
+                    + Thread.currentThread().getName());
+            } else {
+                LOG.trace("The next customer is: " + customer.getName());
+            }
+            return customer;
+        } finally{
+            dequeueLock.unlock();
         }
-        return customer;
+
     }
 
     public int getLineLength() {
@@ -58,7 +68,9 @@ public class ConcurrentCustomerLine extends ConcurrentLinkedQueue<AbstractCustom
             AbstractCustomer person = iterator.next();
             names.add(person.getName());
         }
-        LOG.info("Line: " + names);
+        if (AppConfig.SHOW_CUSTOMERS_IN_LINE) {
+            LOG.info("Line: " + names);
+        }
     }
 
     public boolean isCustomerPatient(AbstractCustomer customer) {
