+++ src/main/java/prescriptionRegistry/PrescriptionRegistry.java
@@ -8,6 +8,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import setup.AppConfig;
 
 
 public class PrescriptionRegistry implements IPrescriptionRegistry {
@@ -38,11 +39,13 @@ public class PrescriptionRegistry implements IPrescriptionRegistry {
 
     public void addPrescription(Patient patient, Prescription prescription)
         throws PersonDoesNotExistException {
-        LOG.trace("Adding prescription " + prescription.getPrescriptionId() + " for patient "
-            + patient.getPatientID());
+        if (AppConfig.SHOW_PRESCRIPTION_REGISTRY_LOGS) {
+            LOG.trace("Adding prescription " + prescription.getPrescriptionId() + " for patient "
+                + patient.getPatientID());
+        }
 
         if (!prescriptionRegistry.containsKey(patient)) {
-            LOG.warn("PersonDoesNotExistException was thrown");
+            if (AppConfig.SHOW_PRESCRIPTION_REGISTRY_LOGS)LOG.warn("PersonDoesNotExistException was thrown");
             throw new PersonDoesNotExistException("The person is not registered in database");
         }
         prescription.setPrescriptionStatus(PrescriptionStatus.PENDING);
