+++ src/main/java/setup/PharmacySetup.java
@@ -11,6 +11,7 @@ import person.PharmacyTechnician;
 import pharmacy.Pharmacy;
 import product.Item;
 import product.Medication;
+import utils.DataProvider;
 
 public class PharmacySetup {
 
@@ -23,13 +24,13 @@ public class PharmacySetup {
     private static final int MAX_NUM_PER_MEDICATION = 500;
 
     private static final IPerformOperation<Pharmacy> hirePharmacyEmployees = (Pharmacy pharmacy) -> {
-        LOG.trace("Start hiring employees");
+       if(AppConfig.SHOW_PHARMACY_SETUP)LOG.trace("Start hiring employees");
         pharmacy.hireEmployee(PHARMACISTS[0]);
         pharmacy.hireEmployee(PHARMACISTS[1]);
         pharmacy.hireEmployee((TECHNICIANS[0]));
         pharmacy.hireEmployee((TECHNICIANS[0])); // hire duplicate
         pharmacy.hireEmployee((TECHNICIANS[1]));
-        LOG.trace("Completed hiring employees");
+        if(AppConfig.SHOW_PHARMACY_SETUP)LOG.trace("Completed hiring employees");
     };
 
     private static final IPerformOperation<Pharmacy> populateInventory = (Pharmacy pharmacy) -> {
@@ -41,7 +42,7 @@ public class PharmacySetup {
         for (Medication medication : DataProvider.predefinedMedications()) {
             productInventory.addProduct(medication, MAX_NUM_PER_MEDICATION);
         }
-        LOG.trace("Completed populating product inventory");
+        if(AppConfig.SHOW_PHARMACY_SETUP)LOG.trace("Completed populating product inventory");
         pharmacy.setProductInventory(productInventory);
     };
 
@@ -67,11 +68,11 @@ public class PharmacySetup {
         for (DayOfWeek day : DayOfWeek.values()) {
             String dayStr = day.name();
             BusinessDay businessDay = BusinessDay.getBusinessDay(dayStr);
-            LOG.trace(businessDay.getDescription());
+            if(AppConfig.SHOW_PHARMACY_SETUP)LOG.trace(businessDay.getDescription());
             if (pharmacy.isOpen(day)) {
-                LOG.trace("Come on in!");
+                if(AppConfig.SHOW_PHARMACY_SETUP)LOG.trace("Come on in!");
             } else {
-                LOG.trace("Come back later!");
+                if(AppConfig.SHOW_PHARMACY_SETUP)LOG.trace("Come back later!");
             }
         }
     }
