+++ src/main/java/utils/CustomerSupplier.java
@@ -1,4 +1,4 @@
-package setup;
+package utils;
 
 import java.util.ArrayList;
 import java.util.List;
