+++ src/main/java/utils/PharmacyTechnicianSupplier.java
@@ -1,4 +1,4 @@
-package setup;
+package utils;
 
 import java.util.ArrayList;
 import java.util.LinkedList;
