+++ src/test/java/inventory/CartTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 import pharmacy.Pharmacy;
-import setup.CartSupplier;
+import utils.CartSupplier;
 import setup.PharmacySetup;
 
 class CartTest {
