+++ src/test/java/register/RegisterTest.java
@@ -3,7 +3,7 @@ package register;
 import static org.junit.jupiter.api.Assertions.*;
 
 import inventory.Cart;
-import setup.DataProvider;
+import utils.DataProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import person.AbstractCustomer;
